backend/process-image/requirements.txt
fal-client==0.4.1
Pillow==10.3.0
requests==2.31.0
